Moved pdb functions out to their own file
