+++ Cargo.toml
@@ -13,9 +13,8 @@ cfg-if = "1.0.0"
 libc = "0.2.82"
 object = "0.26.0"
 pdb = "0.7.0"
-uuid = "0.8"
+uuid = "0.8.2"
 
 [target.'cfg(target_os="macos")'.dependencies]
 core-foundation = "0.9.1"
 core-foundation-sys = "0.8.2"
-uuid = "0.8.2"
